% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{arguments.configs}
\alias{arguments.configs}
\title{Shared documentation for configuration related parameteres}
\arguments{
\item{control}{a list with configurations to be passed to the optimization
subroutines (fit only). Default is \code{control = NULL}. Missing arguments will
receive default values. For details, see \link{fit.control}.}

\item{report}{logical; indicates whether the summary from the fitted model
should be be printed (fit only). Default is \code{report = TRUE}, in which case
\code{info} is automatically set to \code{TRUE}.}

\item{complete}{logical; if FALSE returns only \code{yt}, else returns additional
components (simulation only). Default is \code{complete = FALSE}.}

\item{debug}{logical, if \code{TRUE} the output from FORTRAN is return (for
debugging purposes). Default is \code{debug = FALSE}.}

\item{...}{further arguments passed to the internal functions. See, for
instance, \link{summary.btsr} for details.}
}
\description{
This is the common documentation for arguments related the configurations
for fitting models and printing reports.
}
