% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{BTSR.models}
\alias{BTSR.models}
\title{Table of available model}
\usage{
BTSR.models(do.plot = interactive())
}
\arguments{
\item{do.plot}{logical; if \code{TRUE} returns a plot with the output, otherwise
prints the results in the console.}
}
\value{
\code{NULL} (invisibly). The function is called for its side effects
(printing/plotting).
}
\description{
This function returns the table of available models.
}
