% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btsr.R
\docType{package}
\name{BTSR.Package}
\alias{BTSR.Package}
\alias{_PACKAGE}
\alias{BTSR}
\title{BTSR: Bounded Time Series Regression.}
\description{
The BTSR package provides functions to simulate, estimate and forecast a
wide range of regression based dynamic models for bounded time series. The
package covers the most commonly applied models in the literature.
The package's main calculations are done in FORTRAN, which translates into
very fast algorithms.
}
\author{
Taiane Schaedler Prass \email{taianeprass@gmail.com}
}
\keyword{internal}
